+++ src/containers/sidebars/aoi-sidebar/styles.module.scss
@@ -229,3 +229,8 @@ $sidebar-content-top: 180px;
     }
   }
 }
+
+.mapLayersContainer {
+  @extend %verticalScrollbar;
+  height: 100%;
+}
