+++ src/containers/sidebars/tabs-sidebar/index.js
@@ -1,6 +1,8 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 import { connect } from 'react-redux';
 import aoisActions from 'redux_modules/aois';
+
+import { BASE_LAYERS } from 'constants/aois';
 
 import Component from './component';
 
@@ -9,14 +11,22 @@ const actions = {
 };
 
 function TabsSidebarContainer(props) {
-  const { setSidebarTabActive, onTabClick } = props;
+  const { activeLayers, setSidebarTabActive, onTabClick } = props;
 
   const saveSidebarTab = (selectedTab) => {
     setSidebarTabActive(selectedTab);
   };
 
+  const categoryActiveLayersCounter = useMemo(() => {
+    return activeLayers.map((al, i) => {
+      if (!BASE_LAYERS[i]) return al;
+      return null;
+    }).filter((i) => i !== null);
+  }, [activeLayers]);
+
   return (
     <Component
+      categoryActiveLayersCounter={categoryActiveLayersCounter}
       saveSidebarTab={saveSidebarTab}
       onTabClick={onTabClick}
       {...props}
